+++ packages/components/src/components/combobox/combobox.tsx
@@ -248,7 +248,7 @@ export class Combobox
 
   private get effectiveFilterProps(): string[] {
     if (!this.filterProps) {
-      return ["description", "label", "metadata", "shortHeading", "textLabel"];
+      return ["description", "label", "metadata", "shortHeading"];
     }
 
     return this.filterProps.filter((prop) => prop !== "el");
@@ -1375,7 +1375,6 @@ export class Combobox
       label: item.heading,
       metadata: item.metadata,
       shortHeading: item.shortHeading,
-      textLabel: item.textLabel,
       el: item, // used for matching items to data
     }));
   }
@@ -1407,7 +1406,7 @@ export class Combobox
   }
 
   private addCustomChip(value: string, focus?: boolean): void {
-    const existingItem = this.items.find((el) => (el.heading || el.textLabel) === value);
+    const existingItem = this.items.find((el) => el.heading === value);
     if (existingItem) {
       this.toggleSelection(existingItem, true);
     } else {
@@ -1791,7 +1790,7 @@ export class Combobox
         ariaLabel: item.label,
         ariaSelected: item.selected,
         id: `${IDS.item(item.guid)}`,
-        textContent: item.heading || item.textLabel,
+        textContent: item.heading,
       }),
     );
 
@@ -1833,6 +1832,7 @@ export class Combobox
               this.selectionMode !== "single-persist" && (
                 <calcite-combobox-item
                   class={CSS.selectAll}
+                  heading={messages.selectAll}
                   id={`${this.guid}-select-all-enabled-interactive`}
                   indeterminate={this.indeterminate}
                   label={messages.selectAll}
@@ -1840,7 +1840,6 @@ export class Combobox
                   scale={scale}
                   selected={this.allSelected}
                   tabIndex="-1"
-                  text-label={messages.selectAll}
                   value="select-all"
                 />
               )}