+++ packages/components/src/components/flow-item/flow-item.e2e.ts
@@ -293,10 +293,6 @@ describe("calcite-flow-item", () => {
       },
       "--calcite-flow-border-color": [
         {
-          shadowSelector: `.${CSS.backButton}`,
-          targetProp: "borderColor",
-        },
-        {
           shadowSelector: "calcite-panel",
           targetProp: "--calcite-panel-border-color",
         },