+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-design-tokens@4.0.0-next.11...@esri/calcite-design-tokens@4.0.0-next.12) (2025-12-10)
+
+### ⚠ BREAKING CHANGES
+
+- restore `--calcite-color-focus` in index.(s)css (#13398)
+
+### Bug Fixes
+
+- restore `--calcite-color-focus` in index.(s)css ([#13398](https://github.com/Esri/calcite-design-system/issues/13398)) ([cb63f5f](https://github.com/Esri/calcite-design-system/commit/cb63f5ff5c8754c8eeae6218390636fafb30adb6))
+
 ## [4.0.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-design-tokens@4.0.0-next.10...@esri/calcite-design-tokens@4.0.0-next.11) (2025-12-09)
 
 ### Features